[GUI] move the Mainwindow class in a new file
